Making the `Backbone.sync` override to Firebase actually...work.
When calling `store[method].apply` in `Backbone.Firebase.sync`, the context was being set to `this` which was a reference to the Backbone Model (or Collection) that the sync was triggered from.

In order for the `Backbone.Firebase[method]` methods to access their reference to the Firebase, the context for this call has to be a reference to the actual Firebase store set on the model or collection. Otherwise variables like `this._fbref` and `this._children` are undefined and therefore cause problems.