Simplifying  Backbone.Firebase.Collection constructor
Collections bubble events from models by default, so listening to each model on a collection for add, change, and remove events as well as managing to remove the listeners for those events appropriately is more overhead than is necessary. We can simply listen to the Collection's add/change/remove event proxy, which passes the model as an argument to the callback. We don't even have to listen to the add/remove events for the models, since the 'change' event will only be triggered on models that are part of the collection.